Ktlint refactoring
