+++ app/src/main/kotlin/IT3030/DeepLearning/Config/ActivationFunctionEnum.kt
@@ -1,7 +1,7 @@
 package IT3030.DeepLearning.Config
 
-enum class ActivationFunctionEnum{
-  tanh,
-  linear,
-  softmax,
+enum class ActivationFunctionEnum {
+    tanh,
+    linear,
+    softmax,
 }
