+++ app/src/main/kotlin/IT3030/DeepLearning/Config/WeightRegularizationEnum.kt
@@ -1,5 +1,5 @@
 package IT3030.DeepLearning.Config
 
 enum class WeightRegularizationEnum {
-  glorot
+    glorot
 }
