+++ app/src/test/kotlin/IT3030/DeepLearning/Config/ConfigParserSpec.kt
@@ -1,7 +1,6 @@
 import IT3030.DeepLearning.Config.*
 import com.charleskorn.kaml.Yaml
 import kotlinx.serialization.Serializable
-
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
 import kotlin.test.assertEquals
@@ -9,14 +8,12 @@ import kotlin.test.assertTrue
 
 class ConfigParserSpec : Spek({
 
-
-
     describe("ConfigParser") {
-      assertTrue { true }
+        assertTrue { true }
 
-      describe("It can parse yaml correctly") {
-        // Arrange
-        val configSimple = """
+        describe("It can parse yaml correctly") {
+            // Arrange
+            val configSimple = """
               neuralnet:
                 test: yes!
                 list:
@@ -26,38 +23,41 @@ class ConfigParserSpec : Spek({
                     b: 4
             """.trimIndent()
 
-        @Serializable
-        data class Test(
-          val a: String,
-          val b: String
-        )
-        @Serializable
-        data class Neuralnet(
-          val test: String,
-          val list: List<Test>)
+            @Serializable
+            data class Test(
+                val a: String,
+                val b: String
+            )
+            @Serializable
+            data class Neuralnet(
+                val test: String,
+                val list: List<Test>
+            )
 
-        @Serializable
-        data class SimpleConfig (val neuralnet: Neuralnet)
+            @Serializable
+            data class SimpleConfig(val neuralnet: Neuralnet)
 
+            // Act
+            val parsed = Yaml.default.decodeFromString(SimpleConfig.serializer(), configSimple)
+            println(parsed)
 
-        // Act
-        val parsed = Yaml.default.decodeFromString(SimpleConfig.serializer(), configSimple)
-        println(parsed)
-
-        val expected = SimpleConfig(
-          Neuralnet("yes!", listOf(
-            Test("1", "2"),
-            Test("3", "4")
-          )),
-        )
-        // Assert
-        assertEquals(expected, parsed)
-        assertEquals(expected.neuralnet.test, "yes!")
-        assertEquals(expected.neuralnet.list.size, 2)
-      }
-      describe("It can decode advanced Yaml") {
-        // Arrange
-        val configYml = """
+            val expected = SimpleConfig(
+                Neuralnet(
+                    "yes!",
+                    listOf(
+                        Test("1", "2"),
+                        Test("3", "4")
+                    )
+                ),
+            )
+            // Assert
+            assertEquals(expected, parsed)
+            assertEquals(expected.neuralnet.test, "yes!")
+            assertEquals(expected.neuralnet.list.size, 2)
+        }
+        describe("It can decode advanced Yaml") {
+            // Arrange
+            val configYml = """
                         neuralnet:
                           input:
                             - 20
@@ -79,39 +79,35 @@ class ConfigParserSpec : Spek({
                               - 5
                             weight_regularization: glorot
                             activation_function: softmax
-                        """.trimIndent()
-        println(configYml)
+            """.trimIndent()
+            println(configYml)
 
+            // Act
+            val parsed = Yaml.default.decodeFromString(Config.serializer(), configYml)
+            println("Size")
 
-        //Act
-        val parsed = Yaml.default.decodeFromString(Config.serializer(), configYml)
-        println("Size")
+            println(parsed.neuralnet.hiddenLayers.size)
+            println(parsed)
+            val expected = Config(
+                Neuralnet(
+                    input = listOf(20, 30),
+                    hiddenLayers = listOf(
+                        HiddenLayer(
+                            size = listOf(20, 20),
+                            activation_function = ActivationFunctionEnum.tanh,
+                            weight_regularization = WeightRegularizationEnum.valueOf("glorot")
+                        )
 
-        println(parsed.neuralnet.hiddenLayers.size)
-        println(parsed)
-        val expected = Config(
-          Neuralnet(
-          input = listOf(20, 30),
-          hiddenLayers = listOf(
-            HiddenLayer(
-              size = listOf(20, 20),
-              activation_function = ActivationFunctionEnum.tanh,
-              weight_regularization = WeightRegularizationEnum.valueOf("glorot")
+                    ),
+                    output = Output(
+                        size = listOf(5),
+                        activation_function = ActivationFunctionEnum.softmax,
+                        weight_regularization = WeightRegularizationEnum.valueOf("glorot")
+                    )
+                )
             )
-
-          ),
-            output = Output(
-              size = listOf(5),
-              activation_function = ActivationFunctionEnum.softmax,
-              weight_regularization = WeightRegularizationEnum.valueOf("glorot")
-            )
-        )
-        )
-        // Assert
-        assertEquals(expected.neuralnet.hiddenLayers[0].activation_function.name, "tanh")
-      }
-
-
+            // Assert
+            assertEquals(expected.neuralnet.hiddenLayers[0].activation_function.name, "tanh")
+        }
     }
-
 })
